Fixed 64bit build (thanks to jpv950)

+++ trunk/src/libs/common/Format.cpp
@@ -23,7 +23,6 @@
 //
 
 #include "Format.h"
-#include "../../Types.h"
 
 //! Known type-modifiers. 
 enum Modifiers
@@ -429,9 +428,9 @@ CFormat& CFormat::operator%(void * value)
 		// - Linux:   leading zeros missing 
 		// -> format it as hex
 		if (sizeof (void *) == 8) { // 64 bit
-			SetCurrentField(wxString::Format(wxT("0x%016x"), (uint64) value));
+			SetCurrentField(wxString::Format(wxT("0x%016x"), (uintptr_t) value));
 		} else { // 32 bit
-			SetCurrentField(wxString::Format(wxT("0x%08x"), (uint32) value));
+			SetCurrentField(wxString::Format(wxT("0x%08x"),  (uintptr_t) value));
 		}
 	}
